requirements.txt
behave==1.2.6
selenium
allure-behave
allure-pytest
requests
requests-pkcs12
certifi==2.27.1
urllib3
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pymysql
numpy
pytest_bdd
pandas
webdriver-manager
pytest
allure-python-commons
pyyaml
twilio
pyOpenSSL
pycparser
pyodbc
pyparsing
pypyodbc